created the Who Am I (20 questions) questions object
So when I was coding, it seemed like it was going to be easier to make
the ‘Who Am I?’ style game becuase I wouldn’t have to figure out how to
make dummy answers.

The Who Am I game is semi-implemented in that when you hit ‘go’, a
‘questions’ object is created that contains information about a user.
And then just to demo it, I printed that information to the index.php
file.

The problem is, like you guys were saying at our last meeting, even
with all the information I’m showing about a person, it’s still really
hard to guess who the person is.

Now that I’ve created the ‘Who Am I’ style game, I think I can figure
out how to the Multiple Choice Stalker Book game

BUG: Every time you click the ‘Guess’ button, the page refreshes. How
do I stop it from doing that?