giving up on the multiple choice style game
creating the dummy answers seems like it’s going to be too hard.

instead, i’m going to amend the who am i game to show pictures of
friends along with the clues. One of those friends will be the correct
answer. I feel like this will make it easier to guess the target friend